human-or-ai/requirements.txt
pandas==2.2.3
numpy==2.1.3
scikit-learn==1.5.2
joblib==1.4.2
requests==2.32.3
feedparser==6.0.11
streamlit==1.40.2
pytest==8.3.3
matplotlib==3.9.2
openai==1.57.2
google-generativeai==0.8.3
